+++ src/lib/bio/retry.h
@@ -151,7 +151,7 @@ ssize_t		fr_bio_retry_rewrite(fr_bio_t *bio, fr_bio_retry_entry_t *retry_ctx, co
 
 fr_bio_retry_info_t const *fr_bio_retry_info(fr_bio_t *bio) CC_HINT(nonnull);
 
-size_t		fr_bio_retry_outstanding(fr_bio_t *bio) CC_HINT(nonnull);
+uint64_t       	fr_bio_retry_outstanding(fr_bio_t *bio) CC_HINT(nonnull);
 
 fr_bio_retry_entry_t *fr_bio_retry_item_reserve(fr_bio_t *bio) CC_HINT(nonnull);
